Clean up GroupBy.tsx
